[REFACTOR][CONFIG] Replace dynamic config module with standard class instance
- Remove sys.modules replacement pattern in config.py
- Change Config from types.ModuleType subclass to regular class
- Export config as a standard class instance
- Update all imports to use the new config module structure

This change improves code clarity and follows standard Python patterns
for configuration management.
