+++ examples/load_store.py
@@ -3,7 +3,7 @@ import triton
 import triton.language as tl
 import triton_viz
 from triton_viz.clients import Tracer
-from triton_viz.core import config as cfg
+from triton_viz.core.config import config as cfg
 from triton_viz.core.trace import launches
